% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{acacia_carbon}
\alias{acacia_carbon}
\title{Acacia mangium Willd. | Amazônia}
\usage{
acacia_carbon(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: RR | Compartimento: Total (abaixo+acima) | R2: 0,96 | Syx: 1,73% | Peso: Fresco
DBH range (cm): 8,0 a 300 | Autor: Gomes (2017)
referência: Gomes, J. P. (2017). EQUAÇÕES DE VOLUME E BIOMASSA PARA PLANTIOS DE Acacia mangium Willd. EM ÁREA DE SAVANA, EM RORAIMA. Universidade Federal do Amazonas, 2017.
}
