% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{acacia_mearnsii_de_wild_pampa}
\alias{acacia_mearnsii_de_wild_pampa}
\title{Acacia mangium Willd. | Amazônia}
\usage{
acacia_mearnsii_de_wild_pampa(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: RR | Compartimento: Total (abaixo+acima) | R2: 0,97 | Syx: 1,7% | Peso: Fresco
DBH range (cm): 8,0 a 300 | Autor: Gomes (2017)
referência: GOMES, J. P. Equações de volume e biomassa para plantios de Acacia mangium Willd. em área de savana, em Roraima. Universidade Federal do Amazonas, 2017.
}
