% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{anadenanthera_colubrina_vell_brenan_caatinga}
\alias{anadenanthera_colubrina_vell_brenan_caatinga}
\title{Anadenanthera colubrina (Vell.) Brenan | Caatinga}
\usage{
anadenanthera_colubrina_vell_brenan_caatinga(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: PE | Compartimento: Acima (tronco+galhos+folhas) | R2: 93,45 | Syx: 15,1% | Peso: Fresco
DBH range (cm): 3,0 a 260 | Autor: Abreu et al. (2016)
referência: ABREU, J. C. et al. Ajuste de modelos matemáticos lineares e não lineares para estimativa de biomassa e nutrientes de Anadenanthera colubrina var. cebil no semiárido pernambucano. Scientia Forestalis, v. 44, n. 111, p. 739–750, 2016.
}
