% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{anadenanthera_colubrina_vell_brenan_cerrado}
\alias{anadenanthera_colubrina_vell_brenan_cerrado}
\title{Anadenanthera colubrina (Vell.) Brenan | Cerrado}
\usage{
anadenanthera_colubrina_vell_brenan_cerrado(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: MG | Compartimento: Acima (tronco+galhos+folhas) | R2: 94,99 | Syx: 21,2% | Peso: Seco
DBH range (cm): 10,0 a 400 | Autor: Scolforo et al. (2008)¹
referência: SCOLFORO, J. R. et al. Inventário Florestal de Minas Gerais-Equações de Volume, Peso de Matéria Seca e Carbono para Diferentes Fisionomias da Flora Nativa. Editora UFLA, 2008.
}
