% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{anadenanthera_colubrina_vell_brenan_mata_atlantica}
\alias{anadenanthera_colubrina_vell_brenan_mata_atlantica}
\title{Anadenanthera colubrina (Vell.) Brenan | Mata Atlântica}
\usage{
anadenanthera_colubrina_vell_brenan_mata_atlantica(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: MG | Compartimento: Acima (tronco+galhos+folhas) | R2: 94,50 | Syx: 28,8% | Peso: Seco
DBH range (cm): 10,0 a 400 | Autor: Scolforo et al. (2008)¹
}
