% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{aniba_rosiodora_ducke_amazonia}
\alias{aniba_rosiodora_ducke_amazonia}
\title{Aniba rosiodora Ducke | Amazônia}
\usage{
aniba_rosiodora_ducke_amazonia(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: AM | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,81 | Syx: - | Peso: Fresco
DBH range (cm): 6,0 a 180 | Autor: Krainovic et al. (2017)
referência: KRAINOVIC, P. et al. New allometric equations to support sustainable plantation management of rosewood (Aniba rosaeodora Ducke) in the central Amazon. Forests, v. 8, n. 9, p. 327, 2017.
}
