% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{araucaria_angustifolia_bertol_kuntze_mata_atlantica}
\alias{araucaria_angustifolia_bertol_kuntze_mata_atlantica}
\title{Araucaria angustifolia (Bertol.) Kuntze | Mata Atlântica}
\usage{
araucaria_angustifolia_bertol_kuntze_mata_atlantica(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: PR | Compartimento: Total (abaixo+acima) | R2: 0,97 | Syx: 1,7% | Peso: Seco
DBH range (cm): 20,0 a 440 | Autor: Schumacher et al. (2011)
referência: SCHUMACHER, M. V. et al. Produção de biomassa no corte raso em plantio de Araucaria angustifolia (Bertol.) Kuntze de 27 anos de idade em Quedas do Iguaçu, PR. Ciência Florestal, v. 21, n. 1, p. 53–62, 2011.
}
