% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{araucaria_carbon}
\alias{araucaria_carbon}
\title{Araucaria (Araucaria angustifolia (Bertol.) Kuntze) | Mata Atlântica}
\usage{
araucaria_carbon(dap, h, porc_C)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{h}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa}
}
\value{
Carbono acima do solo [kg]
}
\description{
http://bibliotecaflorestal.ufv.br/bitstream/handle/123456789/15746/Revista_Floresta_v39_n2_p232-237_2009.pdf?sequence=1
}
