% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{arbusto_carbon}
\alias{arbusto_carbon}
\title{Arbustos (genéricos) | Geral}
\usage{
arbusto_carbon(altura_arbusto, crown_diameter, porc_C = 0.47)
}
\arguments{
\item{altura_arbusto}{Altura do arbusto [m]}

\item{crown_diameter}{Diâmetro da copa do arbusto [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Compartimento: Acima (tronco+galhos+folhas) | R2: 0,81 | Syx: - | Peso: Seco
DBH range (cm): - | Autor: Conti et al. (2019)
referência: Conti, G., and Coauthors, 2019: Developing allometric models to predict the individual aboveground biomass of shrubs worldwide. Glob. Ecol. Biogeogr., 28, 961-975, https://doi.org/10.1111/geb.12907.
}
