% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{astronium_urundeuva_m_allemao_engl_caatinga}
\alias{astronium_urundeuva_m_allemao_engl_caatinga}
\title{Astronium urundeuva (M. Allemão) Engl. | Caatinga}
\usage{
astronium_urundeuva_m_allemao_engl_caatinga(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: PE | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,98 | Syx: - | Peso: Seco
DBH range (cm): 2,0 a 500 | Autor: Sampaio e Silva (2005)
referência: SAMPAIO, E. V. S. B.; SILVA, G. C. Biomass equations for Brazilian semiarid caatinga plants. Acta Botanica Brasilica, v. 19, n. 4, p. 935–943, 2005.
}
