% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{ateleia_glazioveana_baill_mata_atlantica}
\alias{ateleia_glazioveana_baill_mata_atlantica}
\title{Ateleia glazioveana Baill. | Mata Atlântica}
\usage{
ateleia_glazioveana_baill_mata_atlantica(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: RS | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,85 | Syx: 31,4% | Peso: Seco
DBH range (cm): 2,0 a 167 | Autor: Fontoura et al. (2017)
referência: FONTOURA, M. R. et al. Modelos alométricos para estimativa de biomassa em área de regeneração natural de Ateleia glazioviana Baill. Floresta, v. 47, n. 4, p. 469–478, 2017.
}
