% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{attalea_phalerata_mart_ex_spreng_outro}
\alias{attalea_phalerata_mart_ex_spreng_outro}
\title{Attalea phalerata Mart. ex Spreng. | Outro}
\usage{
attalea_phalerata_mart_ex_spreng_outro(hf, porc_C = 0.47)
}
\arguments{
\item{hf}{Altura do fuste [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: Outro | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,86 | Syx: - | Peso: Seco
DBH range (cm): 17,0 a 500 | Autor: Goodman et al. (2013)
referência: GOODMAN, R. C. et al. Amazon palm biomass e allometry. Forest Ecology e Management, v. 310, p. 994–1004, 2013.
}
