% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{avicennia_schaueriana_stapf_leechm_ex_moldenke_mata_atlantica}
\alias{avicennia_schaueriana_stapf_leechm_ex_moldenke_mata_atlantica}
\title{Avicennia schaueriana Stapf & Leechm. ex Moldenke | Mata Atlântica}
\usage{
avicennia_schaueriana_stapf_leechm_ex_moldenke_mata_atlantica(
  dap,
  altura,
  porc_C = 0.47
)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: RJ | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,99 | Syx: 17,4% | Peso: Seco
DBH range (cm): 0,7 a 373 | Autor: Estrada et al. (2014)
referência: ESTRADA, G. C. D. et al. Allometric models for aboveground biomass estimation of the mangrove Avicennia schaueriana. Hydrobiologia, v. 734, n. 1, p. 171–185, 2014.
}
