% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{avocado_carbon}
\alias{avocado_carbon}
\title{Avocado (Persea americana Mill.) | Geral}
\usage{
avocado_carbon(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: RR | Compartimento: acima (tronco+galhos+folhas) | R2: 0,86 | Syx: 1,7% | Peso: Fresco
DBH range (cm): 5.3-96.2 | Autor: Shem Kuyah et al. (2024)
referência: Shem Kuyah, Catherine Muthuri, Denis Wakaba, Athanase Rusanganwa Cyamweshi, Paul Kiprotich, Athanase Mukuralinda,
Allometric equations and carbon sequestration potential of mango (Mangifera indica) and avocado (Persea americana) in Kenya,
Trees, Forests and People, Volume 15, 2024, 100467, ISSN 2666-7193, https://doi.org/10.1016/j.tfp.2023.100467.
}
