% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{bambusa_oldhamii_munro_mata_atlantica}
\alias{bambusa_oldhamii_munro_mata_atlantica}
\title{Bambusa oldhamii Munro | Mata Atlântica}
\usage{
bambusa_oldhamii_munro_mata_atlantica(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: SP | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,78 | Syx: 14,6% | Peso: Seco
DBH range (cm): 5,6 a 103 | Autor: Sanquetta et al. (2015a)
referência: SANQUETTA, C. R. et al. Biomassa Individual De Bambusa oldhamii Munro E Bambusa vulgaris Schrad. Ex J.C. Wendl. Cerne, v. 21, n. 1, p. 151–159, 2015a.
}
