% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{bambusa_vulgaris_schrad_ex_j_c_wendl_mata_atlantica}
\alias{bambusa_vulgaris_schrad_ex_j_c_wendl_mata_atlantica}
\title{Bambusa vulgaris Schrad. ex J.C.Wendl. | Mata Atlântica}
\usage{
bambusa_vulgaris_schrad_ex_j_c_wendl_mata_atlantica(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: SP | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,76 | Syx: 16,5% | Peso: Seco
DBH range (cm): 5,7 a 100 | Autor: Sanquetta et al. (2015a)
referência: SANQUETTA, C. R. et al. Biomassa Individual De Bambusa oldhamii Munro E Bambusa vulgaris Schrad. Ex J.C. Wendl. Cerne, v. 21, n. 1, p. 151–159, 2015a.
}
