% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{banana_carbon}
\alias{banana_carbon}
\title{Banana (Musa spp.) | Geral}
\usage{
banana_carbon(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Compartimento: Acima (tronco+galhos+folhas) | R2: - | Syx: - | Peso: Seco
DBH range (cm): - | Autor: Corrêa (2013)
referência: Corrêa, A., 2013: AGROFLORESTA, E C O L O G I A e SOCIEDADE.
}
