% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{bellucia_spp_amazonia}
\alias{bellucia_spp_amazonia}
\title{Bellucia spp. | Amazônia}
\usage{
bellucia_spp_amazonia(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: AM | Compartimento: Acima (tronco+galhos+folhas) | R2: 1,00 | Syx: - | Peso: Seco
DBH range (cm): 1,2 a 268 | Autor: Nelson et al. (1999)
referência: NELSON, B. W. et al. Allometric regressions for improved estimate of secondary forest biomass in the central Amazon. Forest Ecology e Management, v. 117, p. 149–167, 1999.
}
