% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{bertholletia_excelsa_bonpl_amazonia}
\alias{bertholletia_excelsa_bonpl_amazonia}
\title{Bertholletia excelsa Bonpl. | Amazônia}
\usage{
bertholletia_excelsa_bonpl_amazonia(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: AM | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,99 | Syx: 13,8% | Peso: Seco
DBH range (cm): 6,0 a 180 | Autor: Lima (2011)
referência: LIMA, R. M. B. Estimativa da biomassa aérea de castanha-do-Brasil (Bertholletia excelsa Humb. e Bonpl.) em plantios homogêneos no estado do Amazonas. Embrapa Amazônia Ocidental, n. 1, p. 57–61, 2011.
}
