% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{cacau_carbon}
\alias{cacau_carbon}
\title{Cacau (Theobroma cacao L.) | Geral}
\usage{
cacau_carbon(altura, diam_30cm, porc_C = 0.47)
}
\arguments{
\item{altura}{Altura da árvore [m]}

\item{diam_30cm}{Diâmetro da altura de 30 cm [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Compartimento: Acima e abaixo (tronco+galhos+folhas+raizes) | R2: - | Syx: - | Peso: Seco
DBH range (cm): - | Autor: Somarriba et al. (2013)
referência: Somarriba, E., and Coauthors, 2013: Carbon stocks and cocoa yields in agroforestry systems of Central America. Agric. Ecosyst. Environ., 173, 46-57, https://doi.org/10.1016/j.agee.2013.04.013.
}
