% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{palmeira_carbon}
\alias{palmeira_carbon}
\alias{general_carbon}
\alias{coffee_carbon}
\alias{banana_carbon}
\alias{jussara_carbon}
\alias{arbusto_carbon}
\alias{cacau_carbon}
\alias{euterpe_oleracea_carbon}
\alias{euterpe_precatoria_carbon}
\alias{citrus_carbon}
\alias{jaqueira_carbon}
\alias{acacia_carbon}
\alias{crescimento_lento_carbon}
\alias{crescimento_medio_carbon}
\alias{crescimento_rapido_carbon}
\alias{pupunha_carbon}
\alias{castanheira_carbon}
\alias{coco_carbon}
\alias{eucalipto_carbon}
\alias{teca_carbon}
\alias{cupuacu_carbon}
\alias{general_carbon_higuchi}
\alias{general_carbon_mixed_atlantic_forest_abg}
\alias{general_carbon_mixed_atlantic_forest_bgb}
\alias{avocado_carbon}
\alias{general_dense_atlantic_forest_carbon}
\alias{araucaria_carbon}
\alias{cedro_australiano_carbon}
\alias{acacia_mearnsii_de_wild_pampa}
\alias{aniba_rosiodora_ducke_amazonia}
\alias{bellucia_spp_amazonia}
\alias{bertholletia_excelsa_bonpl_amazonia}
\alias{cecropia_sciadophylla_mart_amazonia}
\alias{cedrela_odorata_l_amazonia}
\alias{ceiba_speciosa_a_st_hil_ravenna_amazonia}
\alias{croton_matourensis_aubl_amazonia}
\alias{goupia_glabra_aubl_amazonia}
\alias{guadua_weberbaueri_pilg_amazonia}
\alias{handroanthus_chrysotrichus_mart_ex_dc_mattos_amazonia}
\alias{laetia_procera_poepp_eichler_amazonia}
\alias{schinus_terebinthifolia_raddi_amazonia}
\alias{syzygium_cumini_l_skeels_amazonia}
\alias{tabebuia_aurea_silva_manso_benth_hook_f_ex_s_moore_amazonia}
\alias{varias_amazonia}
\alias{vismia_cayennensis_jacq_pers_amazonia}
\alias{vismia_japurensis_reichardt_amazonia}
\alias{anadenanthera_colubrina_vell_brenan_caatinga}
\alias{anadenanthera_colubrina_var_cebil_griseb_altschul_caatinga}
\alias{aspidosperma_pyrifolium_mart_zucc_caatinga}
\alias{astronium_urundeuva_m_allemao_engl_caatinga}
\alias{bauhinia_cheilantha_bong_steud_caatinga}
\alias{cenostigma_bracteosum_tul_e_gagnon_g_p_lewis_caatinga}
\alias{cenostigma_pyramidale_tul_e_gagnon_g_p_lewis_caatinga}
\alias{cereus_jamacaru_dc_caatinga}
\alias{cnidoscolus_quercifolius_pohl_caatinga}
\alias{croton_heliotropiifolius_kunth_caatinga}
\alias{croton_jacobinensis_baill_caatinga}
\alias{dalbergia_cearensis_ducke_caatinga}
\alias{jatropha_mollissima_pohl_baill_caatinga}
\alias{mimosa_caesalpiniifolia_benth_caatinga}
\alias{mimosa_hostilis_mart_benth_caatinga}
\alias{mimosa_ophthalmocentra_mart_ex_benth_caatinga}
\alias{mimosa_tenuiflora_willd_poir_caatinga}
\alias{monteverdia_rigida_mart_biral_caatinga}
\alias{schinopsis_brasiliensis_engl_caatinga}
\alias{varias_caatinga}
\alias{anadenanthera_colubrina_vell_brenan_cerrado}
\alias{croton_floribundus_spreng_cerrado}
\alias{eremanthus_erythropappus_dc_macleish_cerrado}
\alias{eucalyptus_camaldulensisdehnh_cerrado}
\alias{eucalyptus_grandis_w_hill_cerrado}
\alias{kielmeyera_10oriaceamart_zucc_cerrado}
\alias{kielmeyera_coriacea_mart_zucc_cerrado}
\alias{stryphnodendron_adstringens_mart_coville_cerrado}
\alias{stryphnodendron_adstringens_mart_coville_cerrado_v2}
\alias{varias_cerrado}
\alias{anadenanthera_colubrina_vell_brenan_mata_atlantica}
\alias{araucaria_angustifolia_bertol_kuntze_mata_atlantica}
\alias{ateleia_glazioveana_baill_mata_atlantica}
\alias{avicennia_schaueriana_stapf_leechm_ex_moldenke_mata_atlantica}
\alias{bambusa_oldhamii_munro_mata_atlantica}
\alias{bambusa_vulgaris_schrad_ex_j_c_wendl_mata_atlantica}
\alias{cecropia_glaziovii_snethl_mata_atlantica}
\alias{cryptomeria_japonica_thunb_ex_l_f_d_don_mata_atlantica}
\alias{cyathea_delgadii_sternb_mata_atlantica}
\alias{euterpe_edulis_mart_mata_atlantica}
\alias{guadua_chacoensis_rojas_londo_o_mata_atlantica}
\alias{guadua_chacoensis_rojas_londo_o_mata_atlantica_v2}
\alias{laguncularia_racemosa_l_c_f_gaertn_mata_atlantica}
\alias{leucaena_leucocephala_lam_de_wit_mata_atlantica}
\alias{mimosa_scabrella_benth_mata_atlantica}
\alias{nectandra_grandiflora_nees_mata_atlantica}
\alias{phyllostachys_aurea_riviere_c_riviere_mata_atlantica}
\alias{pimenta_pseudocaryophyllus_gomes_landrum_mata_atlantica}
\alias{pinus_caribaea_var_hondurensis_senecl_w_h_barrett_golfari_mata_atlantica}
\alias{pinus_spp_mata_atlantica}
\alias{pinus_taeda_l_mata_atlantica}
\alias{rhizophora_mangle_l_mata_atlantica}
\alias{tabebuia_cassinoides_lam_dc_mata_atlantica}
\alias{tabebuia_cassinoides_lam_dc_mata_atlantica_v2}
\alias{varias_mata_atlantica}
\alias{attalea_phalerata_mart_ex_spreng_outro}
\alias{euterpe_precatoria_mart_outro}
\alias{iriartea_deltoidea_ruiz_pav_outro}
\alias{mauritia_flexuosa_l_f_outro}
\alias{mauritiella_aculeata_kunth_burret_outro}
\alias{oenocarpus_bataua_mart_outro}
\alias{socratea_exorrhiza_mart_h_wendl_outro}
\alias{varias_outro}
\alias{anadenanthera_colubrina_vell_brenan_pantanal}
\alias{coccoloba_guaranitica_hassler_pantanal}
\alias{diplokeleba_floribunda_n_e_br_pantanal}
\alias{diptychandra_aurantiaca_tul_pantanal}
\alias{licania_minutiflora_sagot_fritsch_pantanal}
\alias{magonia_pubescens_a_st_hil_pantanal}
\alias{phyllostylon_rhamnoides_poiss_taub_pantanal}
\alias{protium_heptaphyllum_aubl_marchand_pantanal}
\alias{terminalia_argentea_mart_zucc_pantanal}
\alias{varias_pantanal}
\title{Set of Carbon equations of ABG}
\usage{
palmeira_carbon(dap, porcentagem_peso_seco = 0.6, porc_C = 0.47)

general_carbon(dap, altura, densidade_madeira, porc_C = 0.47)

coffee_carbon(diam_base, porc_C = 0.47)

banana_carbon(dap, porc_C = 0.47)

jussara_carbon(dap, altura, porc_C = 0.47)

arbusto_carbon(altura_arbusto, crown_diameter, porc_C = 0.47)

cacau_carbon(altura, diam_30cm, porc_C = 0.47)

euterpe_oleracea_carbon(dap)

euterpe_precatoria_carbon(altura, porc_C = 0.47)

citrus_carbon(diam_30cm, porc_C = 0.47)

jaqueira_carbon(dap, porc_C = 0.47)

acacia_carbon(dap, porc_C = 0.47)

crescimento_lento_carbon(dap, altura, porc_C = 0.47)

crescimento_medio_carbon(dap, idade, altura, porc_C = 0.47)

crescimento_rapido_carbon(dap, idade, altura, porc_C = 0.47)

pupunha_carbon(dap, porc_C = 0.47)

castanheira_carbon(dap, porc_C = 0.47)

coco_carbon(altura, porc_C = 0.47)

eucalipto_carbon(dap, altura, porc_C = 0.47)

teca_carbon(dap, porc_C = 0.47)

cupuacu_carbon(diam_30cm, porc_C = 0.47)

general_carbon_higuchi(dap, altura, porc_C = 0.47)

general_carbon_mixed_atlantic_forest_abg(
  dap,
  altura,
  densidade_madeira,
  porc_C = 0.47
)

general_carbon_mixed_atlantic_forest_bgb(
  dap,
  altura,
  densidade_madeira,
  porc_C = 0.47
)

avocado_carbon(dap, porc_C = 0.47)

general_dense_atlantic_forest_carbon(dap, porc_C)

araucaria_carbon(dap, h, porc_C)

cedro_australiano_carbon(dap, h, densidade, porc_C = 0.47)

acacia_mearnsii_de_wild_pampa(dap, porc_C = 0.47)

aniba_rosiodora_ducke_amazonia(dap, altura, porc_C = 0.47)

bellucia_spp_amazonia(dap, altura, porc_C = 0.47)

bertholletia_excelsa_bonpl_amazonia(dap, altura, porc_C = 0.47)

cecropia_sciadophylla_mart_amazonia(dap, porc_C = 0.47)

cedrela_odorata_l_amazonia(dap, porc_C = 0.47)

ceiba_speciosa_a_st_hil_ravenna_amazonia(dap, porc_C = 0.47)

croton_matourensis_aubl_amazonia(dap, porc_C = 0.47)

goupia_glabra_aubl_amazonia(dap, altura, porc_C = 0.47)

guadua_weberbaueri_pilg_amazonia(dap, porc_C = 0.47)

handroanthus_chrysotrichus_mart_ex_dc_mattos_amazonia(
  dap,
  hf,
  altura,
  porc_C = 0.47
)

laetia_procera_poepp_eichler_amazonia(dap, altura, porc_C = 0.47)

schinus_terebinthifolia_raddi_amazonia(dap, porc_C = 0.47)

syzygium_cumini_l_skeels_amazonia(dap, porc_C = 0.47)

tabebuia_aurea_silva_manso_benth_hook_f_ex_s_moore_amazonia(dap, porc_C = 0.47)

varias_amazonia(dap, porc_C = 0.47)

vismia_cayennensis_jacq_pers_amazonia(dap, porc_C = 0.47)

vismia_japurensis_reichardt_amazonia(dap, porc_C = 0.47)

anadenanthera_colubrina_vell_brenan_caatinga(dap, porc_C = 0.47)

anadenanthera_colubrina_var_cebil_griseb_altschul_caatinga(dap, porc_C = 0.47)

aspidosperma_pyrifolium_mart_zucc_caatinga(dap, altura, porc_C = 0.47)

astronium_urundeuva_m_allemao_engl_caatinga(dap, porc_C = 0.47)

bauhinia_cheilantha_bong_steud_caatinga(dap, altura, porc_C = 0.47)

cenostigma_bracteosum_tul_e_gagnon_g_p_lewis_caatinga(
  dap,
  altura,
  porc_C = 0.47
)

cenostigma_pyramidale_tul_e_gagnon_g_p_lewis_caatinga(dap, porc_C = 0.47)

cereus_jamacaru_dc_caatinga(dap, porc_C = 0.47)

cnidoscolus_quercifolius_pohl_caatinga(dap, altura, porc_C = 0.47)

croton_heliotropiifolius_kunth_caatinga(dap, altura, porc_C = 0.47)

croton_jacobinensis_baill_caatinga(dap, porc_C = 0.47)

dalbergia_cearensis_ducke_caatinga(dap, altura, porc_C = 0.47)

jatropha_mollissima_pohl_baill_caatinga(dap, porc_C = 0.47)

mimosa_caesalpiniifolia_benth_caatinga(dap, porc_C = 0.47)

mimosa_hostilis_mart_benth_caatinga(dap, porc_C = 0.47)

mimosa_ophthalmocentra_mart_ex_benth_caatinga(dap, altura, porc_C = 0.47)

mimosa_tenuiflora_willd_poir_caatinga(dap, altura, porc_C = 0.47)

monteverdia_rigida_mart_biral_caatinga(dap, porc_C = 0.47)

schinopsis_brasiliensis_engl_caatinga(dap, porc_C = 0.47)

varias_caatinga(dap, altura, porc_C = 0.47)

anadenanthera_colubrina_vell_brenan_cerrado(dap, altura, porc_C = 0.47)

croton_floribundus_spreng_cerrado(dap, porc_C = 0.47)

eremanthus_erythropappus_dc_macleish_cerrado(dap, altura, porc_C = 0.47)

eucalyptus_camaldulensisdehnh_cerrado(dap, altura, porc_C = 0.47)

eucalyptus_grandis_w_hill_cerrado(dap, altura, porc_C = 0.47)

kielmeyera_10oriaceamart_zucc_cerrado(dap, altura, porc_C = 0.47)

kielmeyera_coriacea_mart_zucc_cerrado(dap, altura, porc_C = 0.47)

stryphnodendron_adstringens_mart_coville_cerrado(dap, altura, porc_C = 0.47)

stryphnodendron_adstringens_mart_coville_cerrado_v2(dap, altura, porc_C = 0.47)

varias_cerrado(dap, altura, porc_C = 0.47)

anadenanthera_colubrina_vell_brenan_mata_atlantica(dap, altura, porc_C = 0.47)

araucaria_angustifolia_bertol_kuntze_mata_atlantica(dap, porc_C = 0.47)

ateleia_glazioveana_baill_mata_atlantica(dap, altura, porc_C = 0.47)

avicennia_schaueriana_stapf_leechm_ex_moldenke_mata_atlantica(
  dap,
  altura,
  porc_C = 0.47
)

bambusa_oldhamii_munro_mata_atlantica(dap, porc_C = 0.47)

bambusa_vulgaris_schrad_ex_j_c_wendl_mata_atlantica(dap, altura, porc_C = 0.47)

cecropia_glaziovii_snethl_mata_atlantica(dap, altura, porc_C = 0.47)

cryptomeria_japonica_thunb_ex_l_f_d_don_mata_atlantica(
  dap,
  altura,
  porc_C = 0.47
)

cyathea_delgadii_sternb_mata_atlantica(dap, altura, porc_C = 0.47)

euterpe_edulis_mart_mata_atlantica(dap, altura, porc_C = 0.47)

guadua_chacoensis_rojas_londo_o_mata_atlantica(dap, altura, porc_C = 0.47)

guadua_chacoensis_rojas_londo_o_mata_atlantica_v2(dap, altura, porc_C = 0.47)

laguncularia_racemosa_l_c_f_gaertn_mata_atlantica(dap, altura, porc_C = 0.47)

leucaena_leucocephala_lam_de_wit_mata_atlantica(
  dap,
  altura,
  hf,
  cap,
  porc_C = 0.47
)

mimosa_scabrella_benth_mata_atlantica(dap, altura, porc_C = 0.47)

nectandra_grandiflora_nees_mata_atlantica(dap, altura, porc_C = 0.47)

phyllostachys_aurea_riviere_c_riviere_mata_atlantica(
  dap,
  altura,
  porc_C = 0.47
)

pimenta_pseudocaryophyllus_gomes_landrum_mata_atlantica(dap, hf, porc_C = 0.47)

pinus_caribaea_var_hondurensis_senecl_w_h_barrett_golfari_mata_atlantica(
  dap,
  altura,
  porc_C = 0.47
)

pinus_spp_mata_atlantica(dap, altura, porc_C = 0.47)

pinus_taeda_l_mata_atlantica(dap, porc_C = 0.47)

rhizophora_mangle_l_mata_atlantica(dap, porc_C = 0.47)

tabebuia_cassinoides_lam_dc_mata_atlantica(dap, porc_C = 0.47)

tabebuia_cassinoides_lam_dc_mata_atlantica_v2(dap, porc_C = 0.47)

varias_mata_atlantica(dap, porc_C = 0.47)

attalea_phalerata_mart_ex_spreng_outro(hf, porc_C = 0.47)

euterpe_precatoria_mart_outro(hf, porc_C = 0.47)

iriartea_deltoidea_ruiz_pav_outro(dap, hf, porc_C = 0.47)

mauritia_flexuosa_l_f_outro(hf, porc_C = 0.47)

mauritiella_aculeata_kunth_burret_outro(hf, porc_C = 0.47)

oenocarpus_bataua_mart_outro(hf, porc_C = 0.47)

socratea_exorrhiza_mart_h_wendl_outro(dap, hf, porc_C = 0.47)

varias_outro(dap, altura, porc_C = 0.47)

acacia_mearnsii_de_wild_pampa(dap, porc_C = 0.47)

anadenanthera_colubrina_vell_brenan_pantanal(dap, porc_C = 0.47)

coccoloba_guaranitica_hassler_pantanal(dap, porc_C = 0.47)

diplokeleba_floribunda_n_e_br_pantanal(dap, porc_C = 0.47)

diptychandra_aurantiaca_tul_pantanal(dap, porc_C = 0.47)

licania_minutiflora_sagot_fritsch_pantanal(dap, porc_C = 0.47)

magonia_pubescens_a_st_hil_pantanal(dap, porc_C = 0.47)

phyllostylon_rhamnoides_poiss_taub_pantanal(dap, porc_C = 0.47)

protium_heptaphyllum_aubl_marchand_pantanal(dap, porc_C = 0.47)

terminalia_argentea_mart_zucc_pantanal(dap, porc_C = 0.47)

varias_pantanal(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da arvore [cm]}

\item{porc_C}{from IPCC Porcentagem da biomassa referente ao carbono (C)/100 = 0.47}

\item{altura}{Altura da árvore [m]}

\item{densidade_madeira}{Densidade da madeira na equação geral [g/cm^3]}

\item{diam_base}{Diametro da base do café [cm]}

\item{altura_arbusto}{Altura do arbusto [m]}

\item{crown_diameter}{Dimetro da copa do arbusto [m]}

\item{diam_30cm}{Diâmetro da altura de 30 cm [cm]}

\item{idade}{Idade da árvore [anos]}
}
\value{
The output returns the amount of Carbon in each plant [kg]
}
\description{
These functions calculate the amount of Carbon in each plant in kg
per individual.
}
\note{
{

Note que o dado da Euterpe oleracea nos dá direto os kg de carbono (C) de cada planta, nao necessitando usar porc_C.

References

Velasco, 2009: Estimativa de sequestro de carbono em mata ciliar: projeto POMAR, São Paulo (SP) / Estimate of carbon sequestration in riparian forests: the POMAR project case, São Paulo (SP). Ambiência, 5, 135-141.

Vieira, S. A., Alves, L. F., Aidar, M., Araújo, L. S., Baker, T., Batista, J. L. F., Campos, M. C., Camargo, P. B., Chave, J., Delitti, W. B. C., Higuchi, N., Honorio, E., Joly, C. A., Keller, M., Martinelli, L. A., de Mattos, E. A., Metzker, T., Phillips, O., dos Santos, F. A. M., … Trumbore, S. E. (2008). Estimation of biomass and carbon stocks: The case of the Atlantic Forest. Biota Neotropica, 8(2), 21-29. https://doi.org/10.1590/S1676-06032008000200001

Alves, L. F., S. A. Vieira, M. A. Scaranello, P. B. Camargo, F. A. M. Santos, C. A. Joly, and L. A. Martinelli, 2010: Forest structure and live aboveground biomass variation along an elevational gradient of tropical Atlantic moist forest (Brazil). For. Ecol. Manage., 260, 679-691, https://doi.org/10.1016/j.foreco.2010.05.023.

Chave, J., and Coauthors, 2005: Tree allometry and improved estimation of carbon stocks and balance in tropical forests. Oecologia, 145, 87-99, https://doi.org/10.1007/s00442-005-0100-x.

Meireles, 2018: CARACTERIZAÇÃO DO ESTOQUE DE CARBONO E EQUAÇÕES ALOMÉTRICAS PARA ESTIMATIVA DE BIOMASSA DE CAFEEIROS ASSOCIADOS À GREVÍLEAS. 0-155.

Corrêa, A., 2013: AGROFLORESTA, E C O L O G I A e SOCIEDADE.

Conti, G., and Coauthors, 2019: Developing allometric models to predict the individual aboveground biomass of shrubs worldwide. Glob. Ecol. Biogeogr., 28, 961-975, https://doi.org/10.1111/geb.12907.

IPCC (Intergovernmental Panel on Climate Change). Forest Lands. In Intergovernmental Panel on Climate Change Guidelines for National Greenhouse Gas Inventories; Institute for Global Environmental Strategies (IGES): Hayama, Japan, 2006; p. 83. Available online: https://www.ipcc-nggip.iges.or.jp/public/2006gl/ (accessed on 22 March 2022).

de Miranda, D. L. C., Sanquetta, C. R., Costa, L. G. da S., & Corte, A. P. D. (2012). Biomassa e carbono em Euterpe oleracea Mart., na Ilha do Marajó - PA. Floresta e Ambiente, 19(3), 336-343. https://doi.org/10.4322/floram.2012.039
Somarriba, E., and Coauthors, 2013: Carbon stocks and cocoa yields in agroforestry systems of Central America. Agric. Ecosyst. Environ., 173, 46-57, https://doi.org/10.1016/j.agee.2013.04.013.

Allometric equations for estimating biomass of euterpe precatoria, the most abundant palm species in the Amazon, da Silva, 2015;

Ètz Schroth, G., Agra, S. D., Geraldes Teixeira, W., Haag, D., & Lieberei, R. (2002). Conversion of secondary forest into agroforestry and monoculture plantations in Amazonia: consequences for biomass, litter and soil carbon stocks after 7 years. In Forest Ecology and Management (Vol. 163).

Saha, C., Mahmood, H., Nayan, S. N. S., Siddique, M. R. H., Abdullah, S. M. R., Islam, S. M. Z., Iqbal, M. Z., & Akhter, M. (2021). Allometric biomass models for the most abundant fruit tree species of Bangladesh: A Non-destructive approach. Environmental Challenges, 3. https://doi.org/10.1016/j.envc.2021.100047

Gomes, J. P. (2017). EQUAÇÕES DE VOLUME E BIOMASSA PARA PLANTIOS DE Acacia mangium Willd. EM ÁREA DE SAVANA, EM RORAIMA.

Carneiro de Miranda, D. L., Galvão de Melo, A. C., & Sanquetta, C. R. (2011). Equações alométricas para estimativa de biomassa e carbono em árvores de reflorestamentos de restauração (Vol. 35).

Malimbwi, R.E. & E., R. & Eid, Tron & Chamshama, Shabani & O., S.. (2016). Allometric volume and biomass models in Tanzania. 10.13140/RG.2.1.1891.5445.

Araújo, L. S. (2022). TEOR DE CARBONO ESTOCADO EM PLANTIOS Khaya grandifoliola DE ORIGEM SEMINAL E CLONAL. Available in https://repositorio.uemasul.edu.br/server/api/core/bitstreams/9d4e1ff7-dc96-4062-a7c7-dc52a12a6f7f/content

Higuchi. (1998). Biomassa da parte aérea da vegetação da floresta tropical Úmida de Terra-Firme da Amazônia Brasileira. Acta Amazonica, 28 (2), 153-166. http://www.scielo.br/pdf/aa/v28n2/1809-4392-aa-28-2-0153.pdf

}
}
\examples{
arbusto_carbon(100, 0.30)

}
