% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{cecropia_glaziovii_snethl_mata_atlantica}
\alias{cecropia_glaziovii_snethl_mata_atlantica}
\title{Cecropia glaziovii Snethl. | Mata Atlântica}
\usage{
cecropia_glaziovii_snethl_mata_atlantica(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: SC | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,98 | Syx: 7,7 kg | Peso: Seco
DBH range (cm): 7,7 a 339 | Autor: Uller et al. (2021)²
referência: ULLER, H. F. et al. Biomass models for three species with different growth forms e geographic distribution in the Brazilian Atlantic forest. Canadian Journal of Forest Research, v. 51, n. 10, p. 1419–1431, 2021.
}
