% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{cecropia_sciadophylla_mart_amazonia}
\alias{cecropia_sciadophylla_mart_amazonia}
\title{Cecropia sciadophylla Mart. | Amazônia}
\usage{
cecropia_sciadophylla_mart_amazonia(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: AM | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,98 | Syx: - | Peso: Seco
DBH range (cm): 5,1 a 382 | Autor: Nelson et al. (1999)
referência: NELSON, B. W. et al. Allometric regressions for improved estimate of secondary forest biomass in the central Amazon. Forest Ecology e Management, v. 117, p. 149–167, 1999.
}
