% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{cedro_australiano_carbon}
\alias{cedro_australiano_carbon}
\title{Cedro australiano (Toona ciliata M.Roem.) | Geral}
\usage{
cedro_australiano_carbon(dap, h, densidade, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{h}{Altura da árvore [m]}

\item{densidade}{Densidade da madeira [g/cm³]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Ng. Polbina Monsang, Keshav Kumar Upadhyay, Rajdeep Chanda, Rajiv Pandey, Shri Kant Tripathi,
Volumetric tree growth models for aboveground biomass estimation of Pinus kesiya and Toona ciliata in Northeast India,
Ecological Frontiers,
2024,ISSN 2950-5097,
https://doi.org/10.1016/j.ecofro.2024.04.010.
}
