% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{citrus_carbon}
\alias{citrus_carbon}
\title{Citrus spp. | Geral}
\usage{
citrus_carbon(diam_30cm, porc_C = 0.47)
}
\arguments{
\item{diam_30cm}{Diâmetro da altura de 30 cm [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Compartimento: Acima (tronco+galhos+folhas) | R2: - | Syx: - | Peso: Seco
DBH range (cm): - | Autor: Schroth et al. (2002)
referência: Schroth, G., Agra, S. D., Geraldes Teixeira, W., Haag, D., & Lieberei, R. (2002). Conversion of secondary forest into agroforestry and monoculture plantations in Amazonia: consequences for biomass, litter and soil carbon stocks after 7 years. In Forest Ecology and Management (Vol. 163).
}
