% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{cnidoscolus_quercifolius_pohl_caatinga}
\alias{cnidoscolus_quercifolius_pohl_caatinga}
\title{Cnidoscolus quercifolius Pohl | Caatinga}
\usage{
cnidoscolus_quercifolius_pohl_caatinga(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: PE | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,72 | Syx: 17,5% | Peso: Seco
DBH range (cm): 1,9 a 95 | Autor: Dalla Lana et al. (2018)
referência: DALLA LANA, M. et al. Biomass equations for caatinga species. Nativa, v. 6, n. 5, p. 517, 2018.
}
