% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{coco_carbon}
\alias{coco_carbon}
\title{Coco-gigante (Cocos nucifera L.) | Geral}
\usage{
coco_carbon(altura, porc_C = 0.47)
}
\arguments{
\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Compartimento: Acima e abaixo (tronco+galhos+folhas) | R2: 0.78 | Syx: - | Peso: Seco
DBH range (cm): - | Autor: Malimbwi et al. (2016)
referência: Malimbwi, R.E. & E., R. & Eid, Tron & Chamshama, Shabani & O., S.. (2016). Allometric volume and biomass models in Tanzania. 10.13140/RG.2.1.1891.5445.
}
