% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{coffee_carbon}
\alias{coffee_carbon}
\title{Café arábica (Coffea arabica) | Bahia}
\usage{
coffee_carbon(diam_base, porc_C = 0.47)
}
\arguments{
\item{diam_base}{Diâmetro da base do café [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: BA | Compartimento: Acima (tronco+galhos+folhas) | R2: - | Syx: - | Peso: Seco
DBH range (cm): - | Autor: Meireles (2018)
referência: Meireles, 2018: CARACTERIZAÇÃO DO ESTOQUE DE CARBONO E EQUAÇÕES ALOMÉTRICAS PARA ESTIMATIVA DE BIOMASSA DE CAFEEIROS ASSOCIADOS À GREVÍLEAS. 0-155.
}
