% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{crescimento_lento_carbon}
\alias{crescimento_lento_carbon}
\title{Espécies com crescimento lento | Geral}
\usage{
crescimento_lento_carbon(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Compartimento: Acima (tronco+galhos+folhas) | R2: 0,95 | Syx: - | Peso: Seco
DBH range (cm): - | Autor: Carneiro de Miranda et al. (2011)
referência: Carneiro de Miranda, D. L., Galvão de Melo, A. C., & Sanquetta, C. R. (2011). Equações alométricas para estimativa de biomassa e carbono em árvores de reflorestamentos de restauração (Vol. 35).
}
