% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{croton_floribundus_spreng_cerrado}
\alias{croton_floribundus_spreng_cerrado}
\title{Croton floribundus Spreng. | Cerrado}
\usage{
croton_floribundus_spreng_cerrado(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: SP | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,98 | Syx: 20,0% | Peso: Seco
DBH range (cm): 5,0 a 279 | Autor: Nicodemo et al. (2016)
referência: NICODEMO, M. L. F. et al. Modelos alométricos para estimação de biomassa aérea e alocação de biomassa de capixingui (Croton floribundus Spreng.) em um sistema silviagrícola. Revista Árvore, v. 40, n. 2, p. 279–288, 2015.
}
