% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{cryptomeria_japonica_thunb_ex_l_f_d_don_mata_atlantica}
\alias{cryptomeria_japonica_thunb_ex_l_f_d_don_mata_atlantica}
\title{Cryptomeria japonica (Thunb. Ex L.f.) D.Don | Mata Atlântica}
\usage{
cryptomeria_japonica_thunb_ex_l_f_d_don_mata_atlantica(
  dap,
  altura,
  porc_C = 0.47
)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: PR | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,96 | Syx: 11,1% | Peso: Seco
DBH range (cm): 14,7 a 354 | Autor: Coutinho et al. (2017)
referência: COUTINHO, V. M. et al. Equações tradicionais e simultâneas para biomassa de compartimentos aéreos de Cryptomeria japonica (L.F.) D. DON. Brazilian Journal of Biometrics, v. 35, n. 1, p. 58–75, 2017.
}
