% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{cyathea_delgadii_sternb_mata_atlantica}
\alias{cyathea_delgadii_sternb_mata_atlantica}
\title{Cyathea delgadii Sternb. | Mata Atlântica}
\usage{
cyathea_delgadii_sternb_mata_atlantica(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: SC | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,92 | Syx: 3,6 kg | Peso: Seco
DBH range (cm): 6,3 a 195 | Autor: Uller et al. (2021)
referência: ULLER, H. F. et al. Biomass models for three species with different growth forms e geographic distribution in the Brazilian Atlantic forest. Canadian Journal of Forest Research, v. 51, n. 10, p. 1419–1431, 2021.
}
