% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{dalbergia_cearensis_ducke_caatinga}
\alias{dalbergia_cearensis_ducke_caatinga}
\title{Dalbergia cearensis Ducke | Caatinga}
\usage{
dalbergia_cearensis_ducke_caatinga(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: CE | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,67 | Syx: - | Peso: Seco
DBH range (cm): 5,0 a 201 | Autor: Nogueira et al. (2021)
}
