% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{diplokeleba_floribunda_n_e_br_pantanal}
\alias{diplokeleba_floribunda_n_e_br_pantanal}
\title{Diplokeleba floribunda N.E.Br. | Pantanal}
\usage{
diplokeleba_floribunda_n_e_br_pantanal(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: MS | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,98 | Syx: - | Peso: Seco
DBH range (cm): 7,0 a 270 | Autor: Padilha (2011)
referência: PADILHA, D. R. C. Fitossociologia e estimativas da biomassa aérea e de carbono em chaco florestado no Brasil. Universidade Federal de Mato Grosso do Sul, 2011.
}
