% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{diptychandra_aurantiaca_tul_pantanal}
\alias{diptychandra_aurantiaca_tul_pantanal}
\title{Diptychandra aurantiaca Tul. | Pantanal}
\usage{
diptychandra_aurantiaca_tul_pantanal(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: MS | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,99 | Syx: - | Peso: Seco
DBH range (cm): 5,0 a 350 | Autor: Salis et al. (2006)
referência: SALIS, R. et al. Biomassa e carbono em fragmentos florestais do Pantanal Mato-Grossense. Revista Árvore, v. 30, n. 2, p. 315-324, 2006.
}
