% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{eremanthus_erythropappus_dc_macleish_cerrado}
\alias{eremanthus_erythropappus_dc_macleish_cerrado}
\title{Eremanthus erythropappus (DC.) MacLeish | Cerrado}
\usage{
eremanthus_erythropappus_dc_macleish_cerrado(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: MG | Compartimento: Acima (tronco+galhos+folhas) | R2: 98,48 | Syx: 20,7% | Peso: Seco
DBH range (cm): 5,0 a 350 | Autor: Scolforo et al. (2008)¹
referência: SCOLFORO, J. R. et al. Inventário Florestal de Minas Gerais-Equações de Volume, Peso de Matéria Seca e Carbono para Diferentes Fisionomias da Flora Nativa. Editora UFLA, 2008.
}
