% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{eucalipto_carbon}
\alias{eucalipto_carbon}
\title{Eucalipto (Eucalyptus spp.) | Geral}
\usage{
eucalipto_carbon(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Compartimento: Acima (tronco+galhos+folhas) | R2: - | Syx: - | Peso: Seco
DBH range (cm): - | Autor: Senelwa e Sims (1998)
referência: Senelwa, K., & Sims, R. E. H. (1998). Eucalypts for biofuel production in New Zealand. Biomass and Bioenergy, 15(2), 109-119.
}
