% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{eucalyptus_camaldulensisdehnh_cerrado}
\alias{eucalyptus_camaldulensisdehnh_cerrado}
\title{Eucalyptus camaldulensis Dehnh. | Cerrado}
\usage{
eucalyptus_camaldulensisdehnh_cerrado(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: MG | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,97 | Syx: 11,5% | Peso: Seco
DBH range (cm): 5,0 a 200 | Autor: Franco (1996)
referência: FRANCO, E. J. Estudo dos métodos estimativos de volume, biomassa e níveis de produtividade para Eucalyptus camaldulensis. Universidade Federal de Lavras, 1996.
}
