% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{eucalyptus_grandis_w_hill_cerrado}
\alias{eucalyptus_grandis_w_hill_cerrado}
\title{Eucalyptus grandis W.Hill | Cerrado}
\usage{
eucalyptus_grandis_w_hill_cerrado(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: MG | Compartimento: Acima (tronco+galhos+folhas) | R2: - | Syx: 6,7% | Peso: Seco
DBH range (cm): 10,0 a 187 | Autor: Ribeiro et al. (2015)
referência: RIBEIRO, S. C. et al. Biomassa acima e abaixo do solo e estimativas de carbono para um plantio clonal de eucalipto no Sudeste do Brasil. Revista Árvore, v. 39, n. 2, p. 353–363, 2015.
}
