% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{euterpe_edulis_mart_mata_atlantica}
\alias{euterpe_edulis_mart_mata_atlantica}
\title{Euterpe edulis Mart. | Mata Atlântica}
\usage{
euterpe_edulis_mart_mata_atlantica(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: SC | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,84 | Syx: 7,1 kg | Peso: Seco
DBH range (cm): 3,9 a 158 | Autor: Uller et al. (2021)
referência: ULLER, H. F. et al. Biomass models for three species with different growth forms e geographic distribution in the Brazilian Atlantic forest. Canadian Journal of Forest Research, v. 51, n. 10, p. 1419–1431, 2021.
}
