% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{euterpe_oleracea_carbon}
\alias{euterpe_oleracea_carbon}
\title{Euterpe oleracea Mart. (Açaí) | Amazônia}
\usage{
euterpe_oleracea_carbon(dap)
}
\arguments{
\item{dap}{DAP da árvore [cm]}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: PA | Compartimento: Acima (tronco+galhos+folhas) | R2: - | Syx: - | Peso: Seco
DBH range (cm): - | Autor: de Miranda et al. (2012)
referência: de Miranda, D. L. C., Sanquetta, C. R., Costa, L. G. da S., & Corte, A. P. D. (2012). Biomassa e carbono em Euterpe oleracea Mart., na Ilha do Marajó - PA. Floresta e Ambiente, 19(3), 336-343. https://doi.org/10.4322/floram.2012.039
}
