% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{euterpe_precatoria_carbon}
\alias{euterpe_precatoria_carbon}
\title{Euterpe precatoria Mart. | Amazônia}
\usage{
euterpe_precatoria_carbon(altura, porc_C = 0.47)
}
\arguments{
\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: Outro | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,97 | Syx: - | Peso: Seco
DBH range (cm): 12,0 a 190 | Autor: Goodman et al. (2013)
referência: GOODMAN, R. C. et al. Amazon palm biomass e allometry. Forest Ecology e Management, v. 310, p. 994–1004, 2013.
}
