% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalparsetext.R
\name{evalparsetext}
\alias{evalparsetext}
\title{Function for evaluate a string}
\usage{
evalparsetext(suastring)
}
\arguments{
\item{suastring}{Simple character object or a vector of strings (characters).}
}
\value{
The output of the function evaluated in suatring vector.
}
\description{
Function that takes your strings and evaluate them, it is useful to
name variables within a for loop. The sky is the limit for this function.
}
\note{
It is necessary to add <<- instead of <- when using evalparsetext function to
add information to a object (see example).
}
\examples{
rm(outp)
evalparsetext(c("outp <<- 1"))
print(outp)
}
