% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{general_carbon}
\alias{general_carbon}
\title{Espécies em geral com DAP >= 5 cm | Geral}
\usage{
general_carbon(dap, altura, densidade_madeira, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{densidade_madeira}{Densidade da madeira [g/cm³]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Compartimento: Acima (tronco+galhos+folhas) | R2: 0,92 | Syx: - | Peso: Seco
DBH range (cm): >= 5 | Autor: Chave et al. (2005)
referência: Chave, J., and Coauthors, 2005: Tree allometry and improved estimation of carbon stocks and balance in tropical forests. Oecologia, 145, 87-99, https://doi.org/10.1007/s00442-005-0100-x.
}
