% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{general_carbon_higuchi}
\alias{general_carbon_higuchi}
\title{Floresta tropical úmida (geral) | Amazônia}
\usage{
general_carbon_higuchi(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Compartimento: Acima (tronco+galhos+folhas) | R2: 0,95 | Syx: - | Peso: Seco
DBH range (cm): - | Autor: Higuchi (1998)
referência: Higuchi. (1998). Biomassa da parte aérea da vegetação da floresta tropical Úmida de Terra-Firme da Amazônia Brasileira. Acta Amazonica, 28 (2), 153-166. http://www.scielo.br/pdf/aa/v28n2/1809-4392-aa-28-2-0153.pdf
}
