% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{general_carbon_mixed_atlantic_forest_bgb}
\alias{general_carbon_mixed_atlantic_forest_bgb}
\title{Espécies em geral com DAP >= 5 cm em Mata Atlântica restaurada (BGB) | Mata Atlântica}
\usage{
general_carbon_mixed_atlantic_forest_bgb(
  dap,
  altura,
  densidade_madeira,
  porc_C = 0.47
)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{densidade_madeira}{Densidade da madeira [g/cm³]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Nogueira Junior, L. R., Engel, V. L., Parrotta, J. A., de Melo, A. C. G., &
Ré, D. S. (2014). Allometric equations for estimating tree biomass in restored
mixed-species Atlantic Forest stands. Biota Neotropica, 14(2). https://doi.org/10.1590/1676-06032013008413
}
