% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{general_dense_atlantic_forest_carbon}
\alias{general_dense_atlantic_forest_carbon}
\title{Fitofisionomia da Mata Atlântica densa | Mata Atlântica}
\usage{
general_dense_atlantic_forest_carbon(dap, porc_C)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa}
}
\value{
Carbono acima do solo [kg]
}
\description{
tiepolo, 2002 & VIEIRA S 2008 & fonseca et al 2024 Em caso de nao ter dados para a fitofisionomia de mata atlantica, usamos chave 2005
}
