% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{guadua_chacoensis_rojas_londo_o_mata_atlantica}
\alias{guadua_chacoensis_rojas_londo_o_mata_atlantica}
\title{Guadua chacoensis (Rojas) Londoño | Mata Atlântica}
\usage{
guadua_chacoensis_rojas_londo_o_mata_atlantica(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: PR | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,86 | Syx: 20,5% | Peso: Seco
DBH range (cm): 5,6 a 140 | Autor: Mognon et al. (2014)
referência: MOGNON, F. et al. Estimativas de biomassa para plantas de bambu do gênero Guadua. Revista Ceres, v. 61, n. 6, p. 900–906, 2014.
}
