% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{guadua_weberbaueri_pilg_amazonia}
\alias{guadua_weberbaueri_pilg_amazonia}
\title{Guadua weberbaueri Pilg. | Amazônia}
\usage{
guadua_weberbaueri_pilg_amazonia(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: AC | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,81 | Syx: - | Peso: Seco
DBH range (cm): 3,5 a 62 | Autor: Torezan e Silveira (2000)
referência: TOREZAN, J. M. D.; SILVEIRA, M. The biomass of bamboo (Guadua weberbaueri Pilger) in open forest of the southwestern Amazon. Ecotropica, v. 6, p. 71–76, 2000.
}
