% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{handroanthus_chrysotrichus_mart_ex_dc_mattos_amazonia}
\alias{handroanthus_chrysotrichus_mart_ex_dc_mattos_amazonia}
\title{Handroanthus chrysotrichus (Mart. ex DC.) Mattos | Amazônia}
\usage{
handroanthus_chrysotrichus_mart_ex_dc_mattos_amazonia(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: MT | Compartimento: Total (abaixo+acima) | R2: 0,96 | Syx: - | Peso: Seco
DBH range (cm): 5,0 a 12,0 | Autor: Robortella (2010)
referência: ROBORTELLA, H. S. Equações de biomassa e estoques de carbono de seis espécies em plantios mistos no noroeste do Mato Grosso. Instituto Nacional de Pesquisas da Amazônia (INPA), 2010.
}
