% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{iriartea_deltoidea_ruiz_pav_outro}
\alias{iriartea_deltoidea_ruiz_pav_outro}
\title{Iriartea deltoidea Ruiz & Pav. | Outro}
\usage{
iriartea_deltoidea_ruiz_pav_outro(dap, hf, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{hf}{Altura do fuste [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: Outro | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,97 | Syx: - | Peso: Seco
DBH range (cm): 6,0 a 330 | Autor: Goodman et al. (2013)
referência: GOODMAN, R. C. et al. Amazon palm biomass e allometry. Forest Ecology e Management, v. 310, p. 994–1004, 2013.
}
