% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{jaqueira_carbon}
\alias{jaqueira_carbon}
\title{Jaqueira (Artocarpus heterophyllus Lam.) | Geral}
\usage{
jaqueira_carbon(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Compartimento: Acima (tronco+galhos+folhas) | R2: - | Syx: - | Peso: Seco
DBH range (cm): - | Autor: Saha et al. (2021)
referência: Saha, C., Mahmood, H., Nayan, S. N. S., Siddique, M. R. H., Abdullah, S. M. R., Islam, S. M. Z., Iqbal, M. Z., & Akhter, M. (2021). Allometric biomass models for the most abundant fruit tree species of Bangladesh: A Non-destructive approach. Environmental Challenges, 3. https://doi.org/10.1016/j.envc.2021.100047
}
