% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{kielmeyera_coriacea_mart_zucc_cerrado}
\alias{kielmeyera_coriacea_mart_zucc_cerrado}
\title{Kielmeyera coriacea Mart. & Zucc. | Cerrado}
\usage{
kielmeyera_coriacea_mart_zucc_cerrado(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: MG | Compartimento: Acima (tronco+galhos+folhas) | R2: 97,64 | Syx: 12,7% | Peso: Seco
DBH range (cm): 5,0 a 250 | Autor: Scolforo et al. (2008)¹
}
