% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{laguncularia_racemosa_l_c_f_gaertn_mata_atlantica}
\alias{laguncularia_racemosa_l_c_f_gaertn_mata_atlantica}
\title{Laguncularia racemosa (L.) C.F.Gaertn. | Mata Atlântica}
\usage{
laguncularia_racemosa_l_c_f_gaertn_mata_atlantica(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: PE | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,97 | Syx: - | Peso: Seco
DBH range (cm): 2,1 a 178 | Autor: Medeiros e Sampaio (2008)
referência: MEDEIROS, T. C. C.; SAMPAIO, E. V. S. B. Allometry of aboveground biomasses in mangrove species in Itamaracá, Pernambuco, Brazil. Wetlands Ecology e Management, v. 16, n. 4, p. 323–330, 2008.
}
