% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{leucaena_leucocephala_lam_de_wit_mata_atlantica}
\alias{leucaena_leucocephala_lam_de_wit_mata_atlantica}
\title{Leucaena leucocephala (Lam.) de Wit | Mata Atlântica}
\usage{
leucaena_leucocephala_lam_de_wit_mata_atlantica(
  dap,
  altura,
  hf,
  cap,
  porc_C = 0.47
)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{hf}{Altura do fuste [m]}

\item{cap}{Capacidade de carga [kg]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: PR | Compartimento: Total (abaixo+acima) | R2: 0,94 | Syx: - | Peso: Seco
DBH range (cm): 1,3 a 150 | Autor: Aleixo et al. (2008)²
referência: ALEIXO, V. et al. Relações Alométricas para Leucaena leucocephala (Lam.) De Wit. Ciência Florestal, v. 18, n. 3, p. 329–338, 2008.
}
