% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{magonia_pubescens_a_st_hil_pantanal}
\alias{magonia_pubescens_a_st_hil_pantanal}
\title{Magonia pubescens A.St.-Hil. | Pantanal}
\usage{
magonia_pubescens_a_st_hil_pantanal(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: MS | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,99 | Syx: - | Peso: Seco
DBH range (cm): 7,0 a 350 | Autor: Salis et al. (2006)
referência: SALIS, S. M. et al. Estimating the aboveground biomass e wood volume of savanna woodlands in Brazil's Pantanal wetlands based on allometric correlations. Forest Ecology e Management, v. 228, n. 1–3, p. 61–68, 2006.
}
