% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{mauritia_flexuosa_l_f_outro}
\alias{mauritia_flexuosa_l_f_outro}
\title{Mauritia flexuosa L.f. | Outro}
\usage{
mauritia_flexuosa_l_f_outro(hf, porc_C = 0.47)
}
\arguments{
\item{hf}{Altura do fuste [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: Outro | Compartimento: Abaixo (raízes) | R2: 0,93 | Syx: - | Peso: Seco
DBH range (cm): 19,0 a 360 | Autor: Goodman et al. (2013)
referência: GOODMAN, R. C. et al. Amazon palm biomass e allometry. Forest Ecology e Management, v. 310, p. 994–1004, 2013.
}
