% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{mauritiella_aculeata_kunth_burret_outro}
\alias{mauritiella_aculeata_kunth_burret_outro}
\title{Mauritiella aculeata (Kunth) Burret | Outro}
\usage{
mauritiella_aculeata_kunth_burret_outro(hf, porc_C = 0.47)
}
\arguments{
\item{hf}{Altura do fuste [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: Outro | Compartimento: Abaixo (raízes) | R2: 0,95 | Syx: - | Peso: Seco
DBH range (cm): 8,0 a 150 | Autor: Goodman et al. (2013)
referência: GOODMAN, R. C. et al. Amazon palm biomass e allometry. Forest Ecology e Management, v. 310, p. 994–1004, 2013.
}
