% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{mimosa_caesalpiniifolia_benth_caatinga}
\alias{mimosa_caesalpiniifolia_benth_caatinga}
\title{Mimosa caesalpiniifolia Benth. | Caatinga}
\usage{
mimosa_caesalpiniifolia_benth_caatinga(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: RN | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,71 | Syx: 13,7% | Peso: Fresco
DBH range (cm): 5,2 a 82 | Autor: Lucena et al. (2020)
}
