% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{mimosa_scabrella_benth_mata_atlantica}
\alias{mimosa_scabrella_benth_mata_atlantica}
\title{Mimosa scabrella Benth. | Mata Atlântica}
\usage{
mimosa_scabrella_benth_mata_atlantica(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: PR | Compartimento: Total (abaixo+acima) | R2: 0,99 | Syx: 0,2% | Peso: Seco
DBH range (cm): 4,3 a 230 | Autor: Mello et al. (2012)
referência: MELLO, A. A. et al. Individual Biomass e Carbon Equations for Mimosa scabrella Benth. (Bracatinga) in Southern Brazil. Silva Fennica, v. 46, n. 3, p. 333–343, 2012.
}
