% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{nectandra_grandiflora_nees_mata_atlantica}
\alias{nectandra_grandiflora_nees_mata_atlantica}
\title{Nectandra grandiflora Nees | Mata Atlântica}
\usage{
nectandra_grandiflora_nees_mata_atlantica(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: PR | Compartimento: Total (abaixo+acima) | R2: 98,44 | Syx: 13,7% | Peso: Seco
DBH range (cm): 9,0 a 350 | Autor: Barbeiro et al. (2009)
referência: BARBEIRO, L. C. et al. Modelos alométricos para estimativa de biomassa em área de regeneração natural de Nectandra grandiflora Nees. Floresta, v. 39, n. 3, p. 505–514, 2009.
}
