% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{phyllostachys_aurea_riviere_c_riviere_mata_atlantica}
\alias{phyllostachys_aurea_riviere_c_riviere_mata_atlantica}
\title{Phyllostachys aurea Rivière & C. Rivière. | Mata Atlântica}
\usage{
phyllostachys_aurea_riviere_c_riviere_mata_atlantica(
  dap,
  altura,
  porc_C = 0.47
)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: PR | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,65 | Syx: 30,2% | Peso: Seco
DBH range (cm): 1,4 a 46 | Autor: Monastier et al. (2015)
referência: MONASTIER, P. et al. Estimativas de biomassa para Phyllostachys aurea Rivière & C. Rivière. Revista Brasileira de Biometria, v. 33, n. 1, p. 1–10, 2015.
}
