% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{phyllostylon_rhamnoides_poiss_taub_pantanal}
\alias{phyllostylon_rhamnoides_poiss_taub_pantanal}
\title{Phyllostylon rhamnoides (Poiss.) Taub. | Pantanal}
\usage{
phyllostylon_rhamnoides_poiss_taub_pantanal(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: MS | Compartimento: Acima (tronco+galhos+folhas) | R2: 1,00 | Syx: - | Peso: Seco
DBH range (cm): 6,0 a 280 | Autor: Padilha (2011)
referência: PADILHA, D. R. C. Fitossociologia e estimativas da biomassa aérea e de carbono em chaco florestado no Brasil. Universidade Federal de Mato Grosso do Sul, 2011.
}
