% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{pimenta_pseudocaryophyllus_gomes_landrum_mata_atlantica}
\alias{pimenta_pseudocaryophyllus_gomes_landrum_mata_atlantica}
\title{Pimenta pseudocaryophyllus (Gomes) Landrum | Mata Atlântica}
\usage{
pimenta_pseudocaryophyllus_gomes_landrum_mata_atlantica(dap, hf, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{hf}{Altura do fuste [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}

\item{altura}{Altura da árvore [m]}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: PR | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,98 | Syx: 5,7% | Peso: Seco
DBH range (cm): 6,0 a 120 | Autor: Girard (2005)
referência: GIRARD, M. A. Estimativas de biomassa para Pimenta pseudocaryophyllus (Gomes) Landrum. Revista Brasileira de Biometria, v. 23, n. 1, p. 1–8, 2005.
}
