% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{pinus_caribaea_var_hondurensis_senecl_w_h_barrett_golfari_mata_atlantica}
\alias{pinus_caribaea_var_hondurensis_senecl_w_h_barrett_golfari_mata_atlantica}
\title{Pinus caribaea var. hondurensis (Sénécl.) W. H. Barrett & Golfari | Mata Atlântica}
\usage{
pinus_caribaea_var_hondurensis_senecl_w_h_barrett_golfari_mata_atlantica(
  dap,
  altura,
  porc_C = 0.47
)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: BA | Compartimento: Acima (tronco+galhos+folhas) | R2: 84,42 | Syx: 25,6% | Peso: Seco
DBH range (cm): 5,8 a 238 | Autor: Lima et al. (2016)
referência: LIMA, E. M. et al. Estimativas de biomassa para Pinus caribaea var. hondurensis (Sénécl.) W. H. Barrett & Golfari. Revista Brasileira de Biometria, v. 34, n. 1, p. 1–10, 2016.
}
