% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{pinus_spp_mata_atlantica}
\alias{pinus_spp_mata_atlantica}
\title{Pinus spp. | Mata Atlântica}
\usage{
pinus_spp_mata_atlantica(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: PR | Compartimento: Total (abaixo+acima) | R2: 0,97 | Syx: 15,4% | Peso: Seco
DBH range (cm): 3,6 a 359 | Autor: Schikowski et al. (2013)
referência: SCHIKOWSKI, P. et al. Estimativas de biomassa para Pinus spp. em áreas de florestas tropicais. Revista Brasileira de Biometria, v. 31, n. 1, p. 1–10, 2013.
}
