% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{pinus_taeda_l_mata_atlantica}
\alias{pinus_taeda_l_mata_atlantica}
\title{Pinus taeda L. | Mata Atlântica}
\usage{
pinus_taeda_l_mata_atlantica(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: RS | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,98 | Syx: 1,5% | Peso: Seco
DBH range (cm): 18,1 a 508 | Autor: Witschoreck (2008)
referência: WITSCHORECK, P. Estimativas de biomassa para Pinus taeda L. em áreas de florestas tropicais. Revista Brasileira de Biometria, v. 26, n. 1, p. 1–10, 2008.
}
