% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{protium_heptaphyllum_aubl_marchand_pantanal}
\alias{protium_heptaphyllum_aubl_marchand_pantanal}
\title{Protium heptaphyllum (Aubl.) Marchand | Pantanal}
\usage{
protium_heptaphyllum_aubl_marchand_pantanal(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: MS | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,97 | Syx: - | Peso: Seco
DBH range (cm): 8,0 a 360 | Autor: Salis et al. (2006)
referência: SALIS, S. M. et al. Estimating the aboveground biomass e wood volume of savanna woodlands in Brazil's Pantanal wetlands based on allometric correlations. Forest Ecology e Management, v. 228, n. 1–3, p. 61–68, 2006.
}
