% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{rhizophora_mangle_l_mata_atlantica}
\alias{rhizophora_mangle_l_mata_atlantica}
\title{Rhizophora mangle L. | Mata Atlântica}
\usage{
rhizophora_mangle_l_mata_atlantica(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: SE | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,99 | Syx: 27,4% | Peso: Seco
DBH range (cm): 1,0 a 392 | Autor: Santos et al. (2017)
referência: SANTOS, E. et al. Estimativas de biomassa para Rhizophora mangle L. em áreas de florestas tropicais. Revista Brasileira de Biometria, v. 35, n. 1, p. 1–10, 2017.
}
