% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{socratea_exorrhiza_mart_h_wendl_outro}
\alias{socratea_exorrhiza_mart_h_wendl_outro}
\title{Socratea exorrhiza (Mart.) H.Wendl. | Outro}
\usage{
socratea_exorrhiza_mart_h_wendl_outro(dap, hf, porc_C = 0.47)
}
\arguments{
\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: Outro | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,98 | Syx: - | Peso: Seco
DBH range (cm): 4,0 a 240 | Autor: Goodman et al. (2013)
referência: GOODMAN, R. C. et al. Amazon palm biomass e allometry. Forest Ecology e Management, v. 310, p. 994–1004, 2013.
}
