% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{stryphnodendron_adstringens_mart_coville_cerrado_v2}
\alias{stryphnodendron_adstringens_mart_coville_cerrado_v2}
\title{Stryphnodendron adstringens (Mart.) Coville | Cerrado (variant)}
\usage{
stryphnodendron_adstringens_mart_coville_cerrado_v2(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: MG | Compartimento: Acima (tronco+galhos+folhas) | R2: 98,68 | Syx: 10,8% | Peso: Seco
DBH range (cm): 5,0 a 250 | Autor: Scolforo et al. (2008)¹
}
