% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{tabebuia_cassinoides_lam_dc_mata_atlantica}
\alias{tabebuia_cassinoides_lam_dc_mata_atlantica}
\title{Tabebuia cassinoides (Lam.) DC. | Mata Atlântica}
\usage{
tabebuia_cassinoides_lam_dc_mata_atlantica(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: PR | Compartimento: Total (abaixo+acima) | R2: 0,96 | Syx: 7,9% | Peso: Seco
DBH range (cm): 8,0 a 159 | Autor: Melo et al. (2014)
referência: MELO, A. et al. Estimativas de biomassa para Tabebuia cassinoides (Lam.) DC. em áreas de florestas tropicais. Revista Brasileira de Biometria, v. 32, n. 1, p. 1–10, 2014.
}
