% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{teca_carbon}
\alias{teca_carbon}
\title{Teca (Tectona grandis L.f.) | Geral}
\usage{
teca_carbon(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Compartimento: Acima (tronco+galhos+folhas) | R2: - | Syx: - | Peso: Seco
DBH range (cm): - | Autor: Pérez e Kanninen (2003)
referência: Pérez, D., & Kanninen, M. (2003). Aboveground biomass and thinning of Tectona grandis plantations. Forest Ecology and Management, 179(1-3), 529-538.
}
