% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{varias_amazonia}
\alias{varias_amazonia}
\title{Várias | Amazônia}
\usage{
varias_amazonia(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: AM | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,93 | Syx: 5,6% | Peso: Seco
DBH range (cm): 5,0 a 1380 | Autor: Lima et al. (2012)
referência: LIMA, A. J. N. et al. Allometric models for estimating above- e below-ground biomass in Amazonian forests at São Gabriel da Cachoeira in the upper Rio Negro, Brazil. Forest Ecology e Management, v. 277, p. 163–172, 2012.
}
