% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{varias_mata_atlantica}
\alias{varias_mata_atlantica}
\title{Várias | Mata Atlântica}
\usage{
varias_mata_atlantica(dap, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: RS | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,98 | Syx: 8,0% | Peso: Seco
DBH range (cm): 3,2 a 550 | Autor: Vogel et al. (2006)
referência: VOGEL, S. et al. Estimativas de biomassa para Várias em áreas de florestas tropicais. Revista Brasileira de Biometria, v. 30, n. 1, p. 1–10, 2006.
}
