% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{varias_tropical_foresta_costa_rica_outro}
\alias{varias_tropical_foresta_costa_rica_outro}
\title{Várias tropical forest in Costa Rica | Outro}
\usage{
varias_tropical_foresta_costa_rica_outro(dap, altura, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: Outro | Compartimento: Acima (tronco+galhos+folhas) | R2: 0,87 | Syx: 14,0% | Peso: Seco
DBH range (cm): 60,0 a 1050 | Autor: Segura e Kanninen (2005)
referência: SEGURA, M.; KANNINEN, M. Allometric models for tree volume e total aboveground biomass in a tropical humid forest in Costa Rica. Biotropica, v. 37, n. 1, p. 2–8, 2005.
}
