% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wood_product.R
\name{wood_product_carbon}
\alias{wood_product_carbon}
\title{Carbon sequestred in the timber products}
\usage{
wood_product_carbon(
  carbono_biomassa_arvores,
  anos,
  wwf_fraction = 0.24,
  slf_class = 0.12,
  fo = 0.86
)
}
\arguments{
\item{carbono_biomassa_arvores}{O carbono total da especie [kgC or tC]}

\item{anos}{Anos pós corte}

\item{wwf_fraction}{0.24 from Verra}

\item{slf_class}{0.12 from Verra}

\item{fo}{0.86 from Verra}
}
\value{
The output returns the carbon sequestered in the products produced with the timber
}
\description{
These functions calculate the amount of Carbon sequestered after some years after cut the trees
}
