% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{araucaria_carbon}
\alias{araucaria_carbon}
\title{Araucaria (Araucaria angustifolia (Bertol.) Kuntze) | Mata Atlântica}
\usage{
araucaria_carbon(dap, h, porc_C)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{h}{Altura da árvore [m]}

\item{porc_C}{Proporção de carbono na biomassa}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: PR | Compartimento: acima (tronco+galhos+folhas) | R2: 0,86 | Syx: 1,7% | Peso: Fresco
DBH range (cm): 5.3-96.2 | Autor: Balbinot et al. (2008)
referência: Balbinot, W. et al. (2008). AJUSTE DE EQUAÇÕES ALOMÉTRICAS PARA Araucaria angustifolia (Bert.) O. Ktze. UTILIZANDO ANÁLISE DE COMPONENTES PRINCIPAIS CONJUNTAMENTE COM ANÁLISE DE REGRESSÃO
}
\details{
NOTA: Plantas de 29 a 33 anos, por isso nao é boa essa equação para os anos inicais
}
