% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{palmeira_carbon}
\alias{palmeira_carbon}
\alias{general_carbon}
\alias{coffee_carbon}
\alias{banana_carbon}
\alias{jussara_carbon}
\alias{carbono_produto_madeira}
\alias{arbusto_carbon}
\title{Set of Carbon equations}
\usage{
palmeira_carbon(dap_palmeira)

general_carbon(dap_arvere, altura, densidade_madeira)

coffee_carbon(diam_base)

banana_carbon(dap)

jussara_carbon(dap, altura)

carbono_produto_madeira(
  carbono_biomassa_arvores,
  anos,
  wwf_fraction = 0.24,
  slf_class = 0.12,
  fo = 0.86
)

arbusto_carbon(altura_arbusto, crown_diameter)
}
\arguments{
\item{dap_palmeira}{DAP da palmeira [cm]}

\item{dap_arvere}{DAP da árvore na equação geral [cm]}

\item{altura}{Altura da árvore [m]}

\item{densidade_madeira}{Densidade da madeira na equação geral [g/cm^3]}

\item{diam_base}{Diametro da base do café [cm]}

\item{dap}{DAP da arvore [cm]}

\item{carbono_biomassa_arvores}{O carbono total da especie [kgC or tC]}

\item{anos}{Anos pós corte}

\item{wwf_fraction}{0.24 from Verra}

\item{slf_class}{0.12 from Verra}

\item{fo}{0.86 from Verra}

\item{altura_arbusto}{Altura do arbusto [cm]}

\item{crown_diameter}{Dimetro da copa do arbusto [m]}
}
\value{
The output returns the amount of Carbon in each plant or
the carbon sequesteres after cutting trees (if using carbono_produto_madeira())
}
\description{
These functions calculate the amount of Carbon in each plant +
the carbon sequestered after some years after cut the trees
}
\examples{
arbusto_carbon(100, 0.30)
}
