% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{palmeira_carbon}
\alias{palmeira_carbon}
\alias{general_carbon}
\alias{coffee_carbon}
\alias{banana_carbon}
\alias{jussara_carbon}
\alias{arbusto_carbon}
\title{Set of Carbon equations of ABG}
\usage{
palmeira_carbon(dap)

general_carbon(dap, altura, densidade_madeira)

coffee_carbon(diam_base)

banana_carbon(dap)

jussara_carbon(dap, altura)

arbusto_carbon(altura_arbusto, crown_diameter)
}
\arguments{
\item{dap}{DAP da arvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{densidade_madeira}{Densidade da madeira na equação geral [g/cm^3]}

\item{diam_base}{Diametro da base do café [cm]}

\item{altura_arbusto}{Altura do arbusto [cm]}

\item{crown_diameter}{Dimetro da copa do arbusto [m]}
}
\value{
The output returns the amount of Carbon in each plant
}
\description{
These functions calculate the amount of Carbon in each plant +
the carbon sequestered after some years after cut the trees
}
\examples{
arbusto_carbon(100, 0.30)
}
