% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{palmeira_carbon}
\alias{palmeira_carbon}
\alias{general_carbon}
\alias{coffee_carbon}
\alias{banana_carbon}
\alias{jussara_carbon}
\alias{arbusto_carbon}
\title{Set of Carbon equations of ABG}
\usage{
palmeira_carbon(dap)

general_carbon(dap, altura, densidade_madeira)

coffee_carbon(diam_base)

banana_carbon(dap)

jussara_carbon(dap, altura)

arbusto_carbon(altura_arbusto, crown_diameter)
}
\arguments{
\item{dap}{DAP da arvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{densidade_madeira}{Densidade da madeira na equação geral [g/cm^3]}

\item{diam_base}{Diametro da base do café [cm]}

\item{altura_arbusto}{Altura do arbusto [cm]}

\item{crown_diameter}{Dimetro da copa do arbusto [m]}
}
\value{
The output returns the amount of Carbon in each plant
}
\description{
These functions calculate the amount of Carbon in each plant +
the carbon sequestered after some years after cut the trees
}
\note{
References
Velasco, 2009: Estimativa de sequestro de carbono em mata ciliar: projeto POMAR, São Paulo (SP) / Estimate of carbon sequestration in riparian forests: the POMAR project case, São Paulo (SP). Ambiência, 5, 135–141.
Vieira, S. A., and Coauthors, 2008: Estimation of biomass and carbon stocks: The case of the Atlantic Forest. Biota Neotrop., 8, 21–29, https://doi.org/10.1590/S1676-06032008000200001.
Alves, L. F., S. A. Vieira, M. A. Scaranello, P. B. Camargo, F. A. M. Santos, C. A. Joly, and L. A. Martinelli, 2010: Forest structure and live aboveground biomass variation along an elevational gradient of tropical Atlantic moist forest (Brazil). For. Ecol. Manage., 260, 679–691, https://doi.org/10.1016/j.foreco.2010.05.023.
Chave, J., and Coauthors, 2005: Tree allometry and improved estimation of carbon stocks and balance in tropical forests. Oecologia, 145, 87–99, https://doi.org/10.1007/s00442-005-0100-x.
Meireles, 2018: CARACTERIZAÇÃO DO ESTOQUE DE CARBONO E EQUAÇÕES ALOMÉTRICAS PARA ESTIMATIVA DE BIOMASSA DE CAFEEIROS ASSOCIADOS À GREVÍLEAS. 0–155.
Corrêa, A., 2013: AGROFLORESTA, E C O L O G I A e SOCIEDADE.
Conti, G., and Coauthors, 2019: Developing allometric models to predict the individual aboveground biomass of shrubs worldwide. Glob. Ecol. Biogeogr., 28, 961–975, https://doi.org/10.1111/geb.12907.
}
\examples{
arbusto_carbon(100, 0.30)
}
