% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_equations.R
\name{cedro_australiano_carbon}
\alias{cedro_australiano_carbon}
\title{Cedro australiano (Toona ciliata M.Roem.) | Geral}
\usage{
cedro_australiano_carbon(dap, altura, densidade, porc_C = 0.47)
}
\arguments{
\item{dap}{DAP da árvore [cm]}

\item{altura}{Altura da árvore [m]}

\item{densidade}{Densidade da madeira [g/cm³]}

\item{porc_C}{Proporção de carbono na biomassa (padrão: 0.47)}
}
\value{
Carbono acima do solo [kg]
}
\description{
Estado: India | Compartimento: acima (tronco+galhos+folhas) | R2: 0,882 | Syx: 1,7% | Peso: Fresco
DBH range (cm): 11.9-53.03 | Autor: Polbina et al. (2024)
referência: Ng. Polbina Monsang, Keshav Kumar Upadhyay, Rajdeep Chanda, Rajiv Pandey, Shri Kant Tripathi, Volumetric tree growth models for aboveground biomass estimation of Pinus kesiya and Toona ciliata in Northeast India,
Ecological Frontiers, Volume 44, Issue 5, 2024, Pages 958-965, ISSN 2950-5097, https://doi.org/10.1016/j.ecofro.2024.04.010.
}
